Bug: 댓글 입력 시 이미지가 깜빡이는 버그 해결
