+++ src/components/screens/MoimBoardStackScreens/postDetail/PostCommentContainer.tsx
@@ -1,4 +1,4 @@
-import {View, Pressable, FlatList} from 'react-native';
+import {View, Pressable} from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import Toast from 'react-native-toast-message';
 
@@ -199,7 +199,7 @@ const PostCommentContainer = ({
     <Pressable onPress={() => targetCommentId && handleUpdateCommentId(null)}>
       <View className="flex flex-col">
         <View
-          className="flex flex-col border-b-[0.5px] border-gray-200 py-4 px-4"
+          className="flex flex-col border-b-[0.5px] border-gray-200 py-4"
           style={{
             backgroundColor:
               targetCommentId === commentData.commentId
@@ -272,16 +272,15 @@ const PostCommentContainer = ({
             좋아요 {commentData?.likeCount}
           </Typography>
         </View>
-        <FlatList
-          data={commentData?.commentResponseDTOList}
-          renderItem={({item}) => (
+        {commentData?.commentResponseDTOList.map(item => (
+          <View key={item.commentId}>
             <PostRecommentContainer
               moimId={moimId}
               postId={postId}
               recommentData={item}
             />
-          )}
-        />
+          </View>
+        ))}
       </View>
     </Pressable>
   );
